# api/src/main/resources/openapi/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "NextRole API",
    "description": "NextRole API Service",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://api.nextrole.app",
      "description": "Prod"
    },
    {
      "url": "http://localhost:9090",
      "description": "Dev"
    }
  ],
  "paths": {
...
    "/auth/otp": {
      "post": {
        "tags": [
          "user"
        ],
        "summary": "Otp confirmation",
        "operationId": "confirmOtp",
        "requestBody": {
          "required": true,
          "description": "OTP",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OtpConfirmation"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Sign in OTP sent confirmation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionUserResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/resume": {
      "get": {
        "tags": [
          "resume"
        ],
        "summary": "Get resume by ID",
        "operationId": "getResume",
        "parameters": [
          {
            "name": "resume_id",
            "schema": {
              "type": "string"
            },
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Resume Data Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResumeData"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "resume"
        ],
        "summary": "Upload user resume",
        "description": "Upload user resume",
        "operationId": "uploadResume",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResumeUploadRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Generic Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResumeUploadResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "resume"
        ],
        "summary": "Update user resume",
        "description": "Update user resume",
        "operationId": "updateResume",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResumeUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Generic Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponse"
                }
              }
            }
          }
        }
      }
    },
    "/job-post": {
      "post": {
        "tags": [
          "resume"
        ],
        "summary": "Parse a job post",
        "operationId": "parsJobPost",
        "requestBody": {
          "required": true,
          "description": "Job post html payload",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ParseJobPostRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Parsed job post response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobPost"
                }
              }
            }
          }
        }
      }
    },
    "/resume/generate": {
      "post": {
        "tags": [
          "resume"
        ],
        "summary": "Generate a resume given a job post",
        "operationId": "generateResume",
        "requestBody": {
          "required": true,
          "description": "JobPost ID",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateResumeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Generated resume",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeneratedResume"
                }
              }
            }
          }
        }
      }
    },
...
    "/resume/template-id": {
      "put": {
        "tags": [
          "resume"
        ],
        "summary": "Update template ID",
        "description": "Update template ID",
        "operationId": "updateTemplateId",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResumeTemplateUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Generic Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "ResumeData": {
        "type": "object",
        "properties": {
          "data": {
            "type": "string"
          }
        }
      },
      "ResumeTemplateUpdateRequest": {
        "type": "object",
        "properties": {
          "resume_id": {
            "type": "string"
          },
          "new_template_id": {
            "type": "string"
          }
        }
      },
      "ResumeUpdateRequest": {
        "type": "object",
        "properties": {
          "fields": {
            "additionalProperties": true
          }
        }
      },
      "ResumeUploadRequest": {
        "type": "object",
        "properties": {
          "file": {
            "type": "string",
            "format": "base64"
          }
        }
      },
      "ResumeUploadResponse": {
        "type": "object",
        "properties": {
          "resume_id": {
            "type": "string"
          }
        }
      },
      "GenerateResumeRequest": {
        "type": "object",
        "properties": {
          "job_id": {
            "type": "string"
          },
          "instructions": {
            "type": "string"
          }
        }
      },
...
      "GeneratedResume": {
        "type": "object",
        "properties": {
          "basics": {
            "type": "object",
            "properties": {
              "name": {
                "type": "string"
              },
              "occupation": {
                "type": "string",
                "description": "e.g. Web Developer"
              },
              "email": {
                "type": "string",
                "description": "e.g. thomas@gmail.com",
                "format": "email"
              },
              "phone": {
                "type": "string",
                "description": "Phone number"
              },
              "url": {
                "type": "string",
                "description": "URL to personal homepage"
              },
              "summary": {
                "type": "string",
                "description": "Write a short 2-3 sentence biography about yourself"
              },
              "location": {
                "type": "object",
                "properties": {
                  "address": {
                    "type": "string",
                    "description": "To add multiple address lines, use \n. For example, 1234 Glücklichkeit Straße\nHinterhaus 5. Etage li."
                  },
                  "postalCode": {
                    "type": "string"
                  },
                  "city": {
                    "type": "string"
                  },
                  "countryCode": {
                    "type": "string",
                    "description": "code as per ISO-3166-1 ALPHA-2, e.g. US, AU, IN"
                  },
                  "region": {
                    "type": "string",
                    "description": "The general region where you live. Can be a US state, or a province, for instance."
                  }
                }
              },
              "profiles": {
                "type": "array",
                "description": "Specify any number of social networks that you participate in",
                "items": {
                  "type": "object",
                  "properties": {
                    "network": {
                      "type": "string",
                      "description": "e.g. Facebook or Twitter"
                    },
                    "username": {
                      "type": "string",
                      "description": "e.g. neutralthoughts"
                    },
                    "url": {
                      "type": "string",
                      "description": "URL to the online profile"
                    }
                  }
                }
              }
            }
          },
          "work": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "description": "e.g. Facebook"
                },
                "location": {
                  "type": "string",
                  "description": "e.g. Menlo Park, CA"
                },
                "description": {
                  "type": "string",
                  "description": "e.g. Social Media Company"
                },
                "position": {
                  "type": "string",
                  "description": "e.g. Software Engineer"
                },
                "url": {
                  "type": "string",
                  "description": "URL to company website"
                },
                "startDate": {
                  "type": "string"
                },
                "endDate": {
                  "type": "string"
                },
                "summary": {
                  "type": "string",
                  "description": "Give an overview of your responsibilities at the company"
                },
                "highlights": {
                  "type": "array",
                  "description": "Specify multiple accomplishments",
                  "items": {
                    "type": "string",
                    "description": "e.g. Increased profits by 20% from 2011-2012 through viral advertising"
                  }
                }
              }
            }
          },
          "volunteer": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "organization": {
                  "type": "string",
                  "description": "e.g. Facebook"
                },
                "position": {
                  "type": "string",
                  "description": "e.g. Software Engineer"
                },
                "url": {
                  "type": "string",
                  "description": "URL to organization website"
                },
                "startDate": {
                  "type": "string"
                },
                "endDate": {
                  "type": "string"
                },
                "summary": {
                  "type": "string",
                  "description": "Give an overview of your responsibilities at the company"
                },
                "highlights": {
                  "type": "array",
                  "description": "Specify accomplishments and achievements",
                  "items": {
                    "type": "string",
                    "description": "e.g. Increased profits by 20% from 2011-2012 through viral advertising"
                  }
                }
              }
            }
          },
          "education": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "institution": {
                  "type": "string",
                  "description": "e.g. Massachusetts Institute of Technology"
                },
                "url": {
                  "type": "string",
                  "description": "URL to the academic institution's home page"
                },
                "area": {
                  "type": "string",
                  "description": "e.g. Arts"
                },
                "studyType": {
                  "type": "string",
                  "description": "e.g. Bachelor"
                },
                "startDate": {
                  "type": "string"
                },
                "endDate": {
                  "type": "string"
                },
                "score": {
                  "type": "string",
                  "description": "grade point average, e.g. 3.67/4.0"
                },
                "courses": {
                  "type": "array",
                  "description": "List notable courses/subjects",
                  "items": {
                    "type": "string",
                    "description": "e.g. H1302 - Introduction to American history"
                  }
                }
              }
            }
          },
          "awards": {
            "type": "array",
            "description": "Specify any awards you have received throughout your professional career",
            "items": {
              "type": "object",
              "properties": {
                "title": {
                  "type": "string",
                  "description": "e.g. One of the 100 greatest minds of the century"
                },
                "date": {
                  "type": "string"
                },
                "awarder": {
                  "type": "string",
                  "description": "e.g. Time Magazine"
                },
                "summary": {
                  "type": "string",
                  "description": "e.g. Received for my work with Quantum Physics"
                }
              }
            }
          },
          "certificates": {
            "type": "array",
            "description": "Specify any certificates you have received throughout your professional career",
            "items": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "description": "e.g. Certified Kubernetes Administrator"
                },
                "date": {
                  "type": "string"
                },
                "url": {
                  "type": "string",
                  "description": "URL, if relevant"
                },
                "issuer": {
                  "type": "string",
                  "description": "e.g. CNCF"
                }
              }
            }
          },
          "publications": {
            "type": "array",
            "description": "Specify your publications through your career",
            "items": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "description": "e.g. The World Wide Web"
                },
                "publisher": {
                  "type": "string",
                  "description": "e.g. IEEE, Computer Magazine"
                },
                "releaseDate": {
                  "type": "string"
                },
                "url": {
                  "type": "string",
                  "description": "URL to the publication"
                },
                "summary": {
                  "type": "string",
                  "description": "Short summary of publication. e.g. Discussion of the World Wide Web, HTTP, HTML."
                }
              }
            }
          },
          "skills": {
            "type": "array",
            "description": "List out your professional skill-set",
            "items": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "description": "e.g. Web Development"
                },
                "level": {
                  "type": "string",
                  "description": "e.g. Master"
                },
                "keywords": {
                  "type": "array",
                  "description": "List some keywords pertaining to this skill",
                  "items": {
                    "type": "string",
                    "description": "e.g. HTML"
                  }
                }
              }
            }
          },
          "languages": {
            "type": "array",
            "description": "List any other languages you speak",
            "items": {
              "type": "object",
              "properties": {
                "language": {
                  "type": "string",
                  "description": "e.g. English, Spanish"
                },
                "fluency": {
                  "type": "string",
                  "description": "e.g. Fluent, Beginner"
                }
              }
            }
          },
          "interests": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "description": "e.g. Philosophy"
                },
                "keywords": {
                  "type": "array",
                  "items": {
                    "type": "string",
                    "description": "e.g. Friedrich Nietzsche"
                  }
                }
              }
            }
          },
          "projects": {
            "type": "array",
            "description": "Specify career projects",
            "items": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "description": "e.g. The World Wide Web"
                },
                "description": {
                  "type": "string",
                  "description": "Short summary of project. e.g. Collated works of 2017."
                },
                "highlights": {
                  "type": "array",
                  "description": "Specify multiple features",
                  "items": {
                    "type": "string",
                    "description": "e.g. Directs you close but not quite there"
                  }
                },
                "keywords": {
                  "type": "array",
                  "description": "Specify special elements involved",
                  "items": {
                    "type": "string",
                    "description": "e.g. AngularJS"
                  }
                },
                "startDate": {
                  "type": "string"
                },
                "endDate": {
                  "type": "string"
                },
                "url": {
                  "type": "string",
                  "description": "Url to the project"
                },
                "roles": {
                  "type": "array",
                  "description": "Specify your role on this project or in company",
                  "items": {
                    "type": "string",
                    "description": "e.g. Team Lead, Speaker, Writer"
                  }
                },
                "entity": {
                  "type": "string",
                  "description": "Specify the relevant company/entity affiliations e.g. 'greenpeace', 'corporationXYZ'"
                },
                "type": {
                  "type": "string",
                  "description": " e.g. 'volunteering', 'presentation', 'talk', 'application', 'conference'"
                }
              }
            }
          }
        }
      },
...
      "ParseJobPostRequest": {
        "type": "object",
        "properties": {
          "job_board": {
            "type": "string"
          },
          "job_description": {
            "type": "string"
          },
          "job_id": {
            "type": "string"
          }
        }
      },
      "JobPost": {
        "type": "object",
        "properties": {
          "job_id": {
            "type": "string"
          },
          "job_title": {
            "type": "string"
          },
          "company_name": {
            "type": "string"
          },
          "company_info": {
            "type": "string"
          },
          "location": {
            "type": "string"
          },
          "salary": {
            "type": "string"
          },
          "logo_url": {
            "type": "string"
          },
          "job_board": {
            "type": "string"
          },
          "job_description": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "text": {
                  "type": "string"
                },
                "is_match": {
                  "type": "boolean"
                }
              }
            }
          }
        }
      },
      "SessionUserResponse": {
        "type": "object",
        "properties": {
          "session_user": {
            "$ref": "#/components/schemas/SessionUser"
          },
          "error": {
            "type": "string"
          }
        }
      },
      "OtpConfirmation": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "otp": {
            "type": "string"
          }
        }
      },
      "GenericResponse":  {
        "type": "object",
        "properties": {
          "status": {
            "type": "string"
          },
          "value": {
            "type": "boolean"
          },
          "any_value": {
            "type": "object"
          }
        }
      },
...
      "SessionUser":  {
        "type": "object",
        "properties": {
          "user_id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "avatar": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "anonymous": {
            "type": "boolean"
          },
          "token": {
            "type": "string"
          },
          "roles": {
            "type": "object",
            "additionalProperties": {
              "type": "boolean"
            }
          }
        }
      },
...
    }
  }
}